Refactored component imports to use a common index file for easier management. Removed unused react-icons dependency from package.json and package-lock.json. Updated Vite configuration to include alias for src directory.
