+++ package.json
@@ -19,7 +19,6 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-icons": "^5.4.0",
     "react-router-dom": "^6.21.3"
   },
   "devDependencies": {
