+++ src/components/common/index.js
@@ -1,0 +1,9 @@
+/**
+ * Common components index file
+ * 
+ * This file re-exports all common components for easier importing.
+ * Instead of importing from specific files, you can import from this index.
+ */
+
+export { default as LazyImage } from './LazyImage';
+export * from './Icons';
